+++ package.json
@@ -61,18 +61,18 @@
   "license": "Apache-2.0",
   "homepage": "https://arc.codes",
   "dependencies": {
-    "@architect/data": "^2.0.8",
+    "@architect/data": "^2.0.9",
     "@architect/parser": "^1.1.6",
     "@smallwins/validate": "^4.3.0",
-    "aws-sdk": "^2.384.0",
+    "aws-sdk": "^2.387.0",
     "body-parser": "^1.18.3",
     "chalk": "^1.1.3",
     "clear": "^0.1.0",
     "cpr": "^3.0.1",
-    "dynalite": "^2.1.0",
+    "dynalite": "^2.2.0",
     "finalhandler": "^1.1.1",
     "glob": "^7.1.3",
-    "inquirer": "^6.2.0",
+    "inquirer": "^6.2.1",
     "is-domain-name": "^1.0.1",
     "log-update": "^2.3.0",
     "mime-types": "^2.1.20",
@@ -85,20 +85,21 @@
     "send": "^0.16.2",
     "shallow-equal": "^1.0.0",
     "strftime": "^0.10.0",
+    "ws": "^6.1.2",
     "zip-dir": "^1.0.2",
     "zipit": "^1.0.2"
   },
   "devDependencies": {
     "cross-env": "^5.2.0",
-    "eslint": "^5.7.0",
+    "eslint": "^5.12.0",
     "nyc": "^13.1.0",
     "proxyquire": "^2.1.0",
-    "rimraf": "^2.6.2",
-    "sinon": "^7.1.1",
+    "rimraf": "^2.6.3",
+    "sinon": "^7.2.2",
     "tap-nyc": "^1.0.3",
     "tap-spec": "^5.0.0",
-    "tape": "^4.9.1",
+    "tape": "^4.9.2",
     "tape-watch": "^2.3.0",
-    "tiny-json-http": "^7.0.0"
+    "tiny-json-http": "^7.0.2"
   }
 }
